backend/requirements.txt
fastapi
uvicorn
gitpython
langchain
transformers
python-dotenv